[Zombie] Health bars, bullet, camera controller added

+++ Assets/Scripts/Zombie/ZombieData.cs
@@ -6,8 +6,8 @@ public class ZombieData
 {
     [SerializeField, Min(0)] private float _moveSpeed;
     public float MoveSpeed => _moveSpeed;
-    [SerializeField, Min(0)] private float _health;
-    public float Health => _health;
+    [SerializeField, Min(0)] private int _health;
+    public int Health => _health;
     [SerializeField] private RuntimeAnimatorController _animatorController;
     public RuntimeAnimatorController AnimatorController => _animatorController;
 }
